+++ Playground/AttributedPlayground.playground/Contents.swift
@@ -22,11 +22,11 @@ let detailsText = Attributed(attributes: [NSForegroundColorAttributeName: NSColo
 //: Usage once you have an Attributed:
 
 //: Apply the formatting to a String
-let stringWithAttributes = redText.string { "Hi mom!" }
+let stringWithAttributes = redText.toString { "Hi mom!" }
 stringWithAttributes
 
 //: Apply the formatting to a NSAttributedString
-let attributedStringWithAttributes = redText.string { NSAttributedString(string: "The quick brown fox") }
+let attributedStringWithAttributes = redText.toString { NSAttributedString(string: "The quick brown fox") }
 attributedStringWithAttributes
 
 //: Combining Strings
@@ -34,11 +34,11 @@ let multipleStrings = boldText.combine(strings: "Hi mom!", "The quick brown fox"
 multipleStrings
 
 //: Nesting
-let nestedStrings = redText.string { boldText.string { "Header: " } + "here are the details" }
+let nestedStrings = redText.toString { boldText.toString { "Header: " } + "here are the details" }
 nestedStrings
 
 //: More nesting
-let anotherNestedString = redText.string { boldText.string { "Header: " } + detailsText.string { "here are the details" } + blueText.string{ "@" } }
+let anotherNestedString = redText.toString { boldText.toString { "Header: " } + detailsText.toString { "here are the details" } + blueText.toString{ "@" } }
 anotherNestedString
 
 //: Strings and attributed strings can be added together