+++ Sources/Attributed+Foundation.swift
@@ -9,7 +9,39 @@
 import AppKit
 import Foundation
 
+// MARK: typealias for AppKit values
+
 public typealias AttributedColor = NSColor
 public typealias AttributedFont = NSFont
 public let AttributedColorAttributeName: String = NSForegroundColorAttributeName
 public let AttributedFontAttributeName: String = NSFontAttributeName
+
+// MARK: NSColor extension
+
+extension NSColor {
+    
+    /// Create an NSAttributedString from a string while applying a 
+    /// NSColor attribute.
+    ///
+    /// - parameter inner: A String
+    /// - returns: An NSAttributedString that has the attributes applied to it
+    public func toString(inner: String) -> NSAttributedString {
+        let attributes = [NSForegroundColorAttributeName: self]
+        return NSAttributedString(string: inner, attributes: attributes)
+    }    
+}
+
+// MARK: NSFont extension
+
+extension NSFont {
+    
+    /// Create an NSAttributedString from a string while applying a
+    /// NSFont attribute.
+    ///
+    /// - parameter inner: A String
+    /// - returns: An NSAttributedString that has the attributes applied to it
+    public func toString(inner: String) -> NSAttributedString {
+        let attributes = [NSFontAttributeName: self]
+        return NSAttributedString(string: inner, attributes: attributes)
+    }
+}