+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Attributed Changelog
 
+## Version 1.1.0
+
+- Changed the `string` method names to `toString`
+- Added a `toString` extension on NSColor, UIColor, NSFont and UIFont
+
 ## Version 1.0.1
 
 - Added a Changelog